pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "cdmon-acme"
version = "0.1.0"
description = "Issue Let's Encrypt certificates (including wildcard) using cdmon DNS via pycdmon"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Pilipilis" }]
dependencies = [
  "pycdmon>=0.1.0",
  "acme>=2.12.0",
  "josepy>=1.15.0",
  "cryptography>=42.0.0",
  "dnspython>=2.6.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.8", "mypy>=1.11"]

[project.scripts]
cdmon-acme = "cdmon_acme.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cdmon_acme"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
